requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
aiohttp==3.9.1
httpx==0.27.2

aiomysql==0.2.0

python-dotenv==1.0.0
aiofiles==23.2.1

APScheduler==3.10.1

numpy==2.1.1
scikit-learn==1.5.2
pydantic==2.10.1

loguru==0.7.2
yacs==0.1.8